// users-service/src/infrastructure/controllers/customers/dtos/validate-student-account-sunlyte.dto.ts
export class ValidateStudentAccountSunlyteDto {
  responseStatusCode: string;
  responseStatusMessage: string;
  OUT_DATE_CREATED: string;
  OUT_SCHOOL_CODE: string;
  OUT_FIRST_NAME: string;
  OUT_MIDDLE_NAME: string;
  OUT_SURNAME: string;
  OUT_STUDENT_CLASS: string;
  OUT_OUTSTANDING_BALANCE: string;
  OUT_STUDENT_PAYMENT_CODE: string;
  OUT_STUDENT_REGISTRATION_NUMBER: string;
  OUT_DATE_OF_BIRTH: string;
  OUT_SCHOOL_NAME: string;
  AIRTEL_OVA: string;
  MTN_OVA: string;
  MINIMUM_ALLOWED_AMOUNT: string;
  OUT_STUDENT_GENDER: string;
}